typesettings ultrasound program

+++ Smart_EPW_lib/ultrasound.h
@@ -1,18 +1,19 @@
 typedef enum{
-		ULTSOUND_BUSY,
-		ULTSOUND_OK
+    ULTSOUND_BUSY,
+    ULTSOUND_OK
 }ultsound_state_err_t;
 
 
 
-unsigned int Get_CH1Distance();
+unsigned long Get_CH1Distance();
 
-unsigned int  Get_CH2Distance();
-unsigned int  Get_CH3Distance();
+unsigned long Get_CH2Distance();
+unsigned long Get_CH3Distance();
 
-unsigned int  Get_CH4Distance();
+unsigned long Get_CH4Distance();
 
 void ultra_sound_init();
+
